Safari-609.1.11.1/LayoutTests/media/modern-media-controls/buttons-container/buttons-container-buttons-property.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<script src="../resources/media-controls-utils.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

description("Testing <code>ButtonsContainer</code> buttons property.");

window.jsTestIsAsync = true;

const container = new ButtonsContainer({
    leftMargin: 20,
    rightMargin: 20,
    buttonMargin: 10
});

container.children = [new Button, new Button, new Button];

scheduler.frameDidFire = function()
{
    shouldBe("container.children.length", "3");
    finishMediaControlsTest();
};

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
